Move jwt token to utils
