+++ concrete/utils.py
@@ -4,11 +4,13 @@ import ast
 import base64
 import os
 import re
+import time
 from datetime import timedelta
 
 import astor
 import dotenv
 import jwt
+from fastapi import HTTPException
 
 dotenv.load_dotenv(override=True)
 
@@ -113,3 +115,61 @@ def verify_jwt(jwt_token: str, access_token: str) -> dict[str, str]:
     assert payload['iss'] in {'https://accounts.google.com', 'accounts.google.com'}
     assert payload['aud'] == os.environ['GOOGLE_OAUTH_CLIENT_ID']
     return payload
+
+
+# TODO merge with verify_jwt
+class JwtToken:
+    """
+    Represents a JWT token.
+    Manages token expiry and generation.
+    """
+
+    def __init__(
+        self,
+        key_name: str,
+        alg: str = "RS256",
+        expiry_offset: int = 600,
+        iss: str | None = None,
+        aud: str | None = None,
+        nbf: int | None = None,
+        additional_headers: dict = {},
+    ):
+        """
+        additional_headers (dict): Headers additional to {typ: 'JWT', alg: alg}
+        """
+        self.iat = None
+        self.exp = None
+        self.alg = alg
+        self.expiry_offset = expiry_offset
+        self.iss = iss
+        self.aud = aud
+        self.nbf = nbf
+        self.additional_headers = additional_headers
+        self.key_value = os.getenv(key_name)
+        if not self.key_value:
+            raise HTTPException(status_code=500, detail=f"{key_name} is not set")
+
+        self._token: str | None = None
+
+    @property
+    def token(self):
+        if not self._token or self._is_expired():
+            self._generate_jwt()
+        return self._token
+
+    def _is_expired(self):
+        return self.exp is None or time.time() >= self.exp
+
+    def _generate_jwt(self):
+        self.iat = time.time()
+        self.exp = self.iat + self.expiry_offset
+        payload = {
+            'exp': self.iat + self.expiry_offset,
+            'iat': self.iat,
+            'iss': self.iss,
+            'aud': self.aud,
+            'nbf': self.nbf,
+        }
+        payload = {k: v for k, v in payload.items() if v is not None}
+
+        self._token = jwt.encode(payload, self.key_value, algorithm=self.alg, headers=self.additional_headers)
